feat : 유료 강의 결제 기능 구현

+++ src/main/java/nextstep/courses/domain/ChargeStatus.java
@@ -6,7 +6,7 @@ public enum ChargeStatus {
 
     public static final String PRICE_OF_FREE_SESSION_ERROR_MESSAGE = "무료 강의의 가격은 0원 입니다.";
 
-    public void validate(int price) {
+    public void validate(long price) {
         if (isFree() && price > 0) {
             throw new IllegalArgumentException(PRICE_OF_FREE_SESSION_ERROR_MESSAGE);
         }
